moved SRC and PUBLIC files into client folder. they were global. app functioning
